+++ korge/src/jvmMain/kotlin/com/soywiz/korge/testing/KorgeScreenshotTestResults.kt
@@ -1,0 +1,6 @@
+package com.soywiz.korge.testing
+
+data class KorgeScreenshotTestResults(
+    val testMethodName: String,
+    val results: MutableList<KorgeScreenshotTestResult> = mutableListOf()
+)
